requirements-core.txt
# Core dependencies (minimal install, no ML models)
fastmcp>=0.2.0
rank_bm25>=0.2.2
langchain-text-splitters>=0.3.0
qdrant-client>=1.11.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Confluence
atlassian-python-api>=3.41.0
html2text>=2020.1.16

# HTTP & Networking
httpx>=0.27.0
tenacity>=8.2.0
urllib3>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-httpx>=0.40b0
prometheus-client>=0.17.0

